fix: 22억 이상의 상금일 시 나타내는 에러 수정
상금이 int의 최대 값을 넘어가면 생기는 에러 수정
Long으로 변경하엿음.

+++ src/main/java/domain/result/Result.java
@@ -17,8 +17,8 @@ public class Result {
     }
 
     public BigDecimal findEarningsRate(BettingMoney bettingMoney) {
-        int prize = results.keySet().stream()
-                .mapToInt(this::getTotalPrize)
+        long prize = results.keySet().stream()
+                .mapToLong(this::getTotalPrize)
                 .sum();
         return bettingMoney.getEarningRate(prize);
     }
@@ -37,7 +37,7 @@ public class Result {
         return copy;
     }
 
-    private int getTotalPrize(LottoRank lottoRank) {
-        return lottoRank.getPrize() * results.get(lottoRank);
+    private Long getTotalPrize(LottoRank lottoRank) {
+        return Long.valueOf(lottoRank.getPrize()) * results.get(lottoRank);
     }
 }
